[Correção#10] Mudar o tipo de variavel do Ruido para aceitar numeros decimais

+++ Core/Model/Metrics.cs
@@ -2,7 +2,7 @@
     public class Metrics {
         public int MetricsId { get; set; }
         public int BatteryLevel { get; set; }
-        public int? Noise{ get; set; }
+        public decimal? Noise{ get; set; }
         public bool IsProtected { get; set; }
         public bool IsContingency { get; set; } = false;
     }
